sparrow/requirements.txt
aiosmtpd==1.4.6
requests==2.32.3
urllib3==2.2.1
idna==3.7
charset-normalizer==3.3.2
certifi==2024.2.2
schedule==1.2.2
psutil==6.0.0
filelock==3.12.0
smbus2==0.4.1
pyserial==3.5
tqdm==4.66.5
Pillow==10.4.0
numpy==1.26.4
pandas==1.4.4
torch==2.4.0
torchvision==0.19.0
torchaudio==2.4.0
tritonclient[http]==2.26.0
protobuf==4.25.3
soundfile==0.12.1
audioread==3.0.1
pooch==1.8.2
librosa==0.10.2.post1
numba==0.59.1
llvmlite==0.42.0